Do the whole parsing/rewriting more rigorously, which works.
This introduces an exported interface, OffsetFrame, which processes a
whole frame (packet in symphonia parlance) by rewriting its
sample/frame offset and recomputing CRCs (which validate correctly!)

Not all the resulting FLAC files are 100% valid yet: "ERROR, decoded
number of samples is smaller than the total number of samples set in
the STREAMINFO" is what `flac --no-decode-through-errors -t`
reports. But they still play and seek in VLC, and this is already a
lot better than what I had previously!