Always write the correct sample counts.
If the samples that the tracks break on do not align with the frame
boundary, we have to adjust the sample counts in the STREAMINFO
header; to that end, buffer the track's audio stream in memory, get
the correct sample count and then write that in the metadata
block. Then write the entire audio stream.

This is a bit unfortunate, as it means the track boundaries aren't
100% correct to the millisecond. But then, I value near-bit-perfect
reconstruction more. Let's leave it at that.